# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: B Y E C O M  is your ecom. shop solution, where you can sell anything
    online with ease and at an affordable price. We provide an eCommerce platform
    for businesses to sell their products online and connect with customers worldwide.
  title: B Y E C O M  LTD
  version: "1.0"
paths:
...
  /admin/user:
    get:
      description: Showing Users details in admin side
      produces:
      - application/json
      responses: {}
      summary: Users Show
      tags:
      - Admin User
    patch:
      description: Blocking or unblocking User
      parameters:
      - description: name search by id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: User Blocking/Unblocking
      tags:
      - Admin User
...
  /user/login:
    post:
      consumes:
      - multipart/form-data
      description: Logging in user with email and pass
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User pass
        in: formData
        name: pass
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Login User
      tags:
      - User Login&Signup
...
  /user/signup:
    delete:
      description: Logging out user
      produces:
      - application/json
      responses: {}
      summary: Logout User
      tags:
      - User Login&Signup
    post:
      consumes:
      - multipart/form-data
      description: Sigining up user
      parameters:
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User pass
        in: formData
        name: pass
        required: true
        type: string
      - description: User phone
        in: formData
        name: phone
        required: true
        type: string
      - description: User gender
        in: formData
        name: gender
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Signup User
      tags:
      - User Login&Signup
  /user/signup/otp:
    post:
      consumes:
      - multipart/form-data
      description: Verifying the user otp and creating user
      parameters:
      - description: User Otp
        in: formData
        name: otp
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Otp verify
      tags:
      - User Login&Signup
...
  /user/wishlist:
    delete:
      description: Removing product from wishlist
      parameters:
      - description: Product id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Remove Wishlist
      tags:
      - User Wishlist
    get:
      description: Showing products from wishlist
      produces:
      - application/json
      responses: {}
      summary: Show Wishlist
      tags:
      - User Wishlist
    post:
      description: Adding a product into wishlist
      parameters:
      - description: Product Id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Wishlist Add
      tags:
      - User Wishlist
swagger: "2.0"